refactor: split r2 cache tests by behavior
